mcp-server/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
mcp>=1.2.1
httpx>=0.26.0
supabase>=2.3.0
PyGithub>=2.1.1
